test(llm-tests): extend live-turn resilience to the thinking matrix
Main CI's push-only live matrix flaked again on agent_run_with_thinking
(test_thinking_with_tool_call::case_1_anthropic_opus — "Should have called
get_current_time"), the same model-non-determinism class already hardened
in agent_run_basic (#2600/#2601).

Lift the retry helper into the shared llm_test_matrix module so every live
test binary can use it: is_transient_transport_error becomes a pub fn and
run_live_turn! is exported via #[macro_export]. agent_run_basic drops its
local copies and resolves the shared ones.

Route agent_run_with_thinking through it: test_thinking_with_tool_call uses
run_live_turn! (retries transport blips + no-tool-call non-determinism);
test_extended_thinking uses a manual bounded loop that keeps the accepted
attempt's runner (its reasoning-field assertions inspect runner.messages(),
which the macro can't return). Genuine non-transient failures still surface
through each test's assertions. CI runs agent_run_basic,
agent_run_with_thinking, subagent_live_test; the first two are now covered.

Test-only; no product code touched.
