+++ crates/llm-tests/tests/agent_run_with_thinking.rs
@@ -18,7 +18,7 @@ use llm_test_matrix::*;
 use rstest::rstest;
 
 use everruns_core::capabilities::CurrentTimeCapability;
-use everruns_core::in_memory_loop::InMemoryAgenticLoop;
+use everruns_core::in_memory_loop::{InMemoryAgenticLoop, TurnResult};
 use everruns_core::message::{ContentPart, Controls, MessageRole, ReasoningConfig};
 use everruns_core::message_retriever::InputMessage;
 
@@ -33,47 +33,85 @@ use everruns_core::message_retriever::InputMessage;
 #[case::openai_gpt54(OPENAI_GPT54)]
 #[tokio::test]
 async fn test_extended_thinking(#[case] config: ProviderModelConfig) {
-    let Some(model) = config.model() else {
+    if config.model().is_none() {
         eprintln!("Skipping: {} not set", config.label());
         return;
-    };
+    }
 
-    let runner = InMemoryAgenticLoop::builder()
-        .agent_name("Thinking Agent")
-        .system_prompt("You are a helpful assistant.")
-        .model(model)
-        .driver_registry(all_providers_registry())
-        .max_iterations(3)
-        .build()
-        .await
-        .unwrap();
+    // Retry the build+run against transient transport blips and adaptive-thinking
+    // non-determinism, but keep the accepted attempt's `runner` because the
+    // reasoning-field assertions below inspect `runner.messages()` — so this
+    // cannot use the `run_live_turn!` macro (which only returns the result).
+    const MAX_ATTEMPTS: usize = 3;
+    let mut accepted: Option<(InMemoryAgenticLoop, TurnResult)> = None;
+    for attempt in 1..=MAX_ATTEMPTS {
+        let model = config.model().expect("model set (checked above)");
+        let runner = InMemoryAgenticLoop::builder()
+            .agent_name("Thinking Agent")
+            .system_prompt("You are a helpful assistant.")
+            .model(model)
+            .driver_registry(all_providers_registry())
+            .max_iterations(3)
+            .build()
+            .await
+            .unwrap();
 
-    // The prompt must be genuinely hard: on adaptive-thinking models (Claude
-    // Fable 5 / Opus 4.8+, GPT-5.x reasoning) the model decides whether to
-    // think at all, and trivial riddles deterministically skip thinking.
-    // Trailing zeros of 2024! = floor(2024/5) + floor(2024/25) + floor(2024/125)
-    // + floor(2024/625) = 404 + 80 + 16 + 3 = 503.
-    let input = InputMessage {
-        role: MessageRole::User,
-        content: vec![ContentPart::text(
-            "How many trailing zeros does 2024! (factorial) have? Work it out carefully, then state the final count as a plain number.",
-        )],
-        controls: Some(Controls {
-            model_id: None,
-            locale: None,
-            reasoning: Some(ReasoningConfig {
-                effort: Some("high".into()),
+        // The prompt must be genuinely hard: on adaptive-thinking models (Claude
+        // Fable 5 / Opus 4.8+, GPT-5.x reasoning) the model decides whether to
+        // think at all, and trivial riddles deterministically skip thinking.
+        // Trailing zeros of 2024! = floor(2024/5) + floor(2024/25) +
+        // floor(2024/125) + floor(2024/625) = 404 + 80 + 16 + 3 = 503.
+        let input = InputMessage {
+            role: MessageRole::User,
+            content: vec![ContentPart::text(
+                "How many trailing zeros does 2024! (factorial) have? Work it out carefully, then state the final count as a plain number.",
+            )],
+            controls: Some(Controls {
+                model_id: None,
+                locale: None,
+                reasoning: Some(ReasoningConfig {
+                    effort: Some("high".into()),
+                }),
+                error_disclosure: None,
+                hints: None,
             }),
-            error_disclosure: None,
-            hints: None,
-        }),
-        metadata: None,
-        tags: vec![],
-    };
+            metadata: None,
+            tags: vec![],
+        };
 
-    let result = runner.run_turn(input).await.unwrap();
+        let result = runner.run_turn(input).await.unwrap();
 
-    skip_if_quota!(result, config.label());
+        skip_if_quota!(result, config.label());
+
+        if result.success && result.response.contains("503") {
+            accepted = Some((runner, result));
+            break;
+        }
+        if attempt == MAX_ATTEMPTS {
+            // Fall through with the last attempt so the assertions below produce
+            // the precise failure message (success / contains-503 / reasoning).
+            accepted = Some((runner, result));
+            break;
+        }
+        let transient = !result.success
+            && result
+                .error
+                .as_deref()
+                .is_some_and(is_transient_transport_error);
+        eprintln!(
+            "{}: attempt {attempt}/{} extended-thinking not accepted \
+             (success={}, transient_transport={}, contains_503={}, error={:?}); retrying",
+            config.label(),
+            MAX_ATTEMPTS,
+            result.success,
+            transient,
+            result.response.contains("503"),
+            result.error,
+        );
+    }
+
+    let (runner, result) = accepted.expect("retry loop always populates a result");
+
     assert!(result.success, "Turn should succeed: {:?}", result.error);
     assert!(
         result.response.contains("503"),
@@ -135,44 +173,57 @@ async fn test_extended_thinking(#[case] config: ProviderModelConfig) {
 #[case::openai_gpt54(OPENAI_GPT54)]
 #[tokio::test]
 async fn test_thinking_with_tool_call(#[case] config: ProviderModelConfig) {
-    let Some(model) = config.model() else {
+    if config.model().is_none() {
         eprintln!("Skipping: {} not set", config.label());
+        return;
+    }
+
+    // Live models are occasionally non-deterministic about emitting a tool call
+    // for this prompt (the "Should have called get_current_time" assertion has
+    // flaked on main), so retry — also absorbing transient transport blips —
+    // and require the tool to be called on at least one attempt. A non-transient
+    // turn failure still surfaces via the asserts below.
+    let Some(result) = run_live_turn!(
+        config,
+        3,
+        |r: &TurnResult| r.success && r.tool_calls_count > 0,
+        {
+            let model = config.model().expect("model set (checked above)");
+            let runner = InMemoryAgenticLoop::builder()
+                .agent_name("Thinking Time Agent")
+                .system_prompt("Use get_current_time tool when asked about time.")
+                .model(model)
+                .driver_registry(all_providers_registry())
+                .capability(CurrentTimeCapability)
+                .max_iterations(5)
+                .build()
+                .await
+                .unwrap();
+            let input = InputMessage {
+                role: MessageRole::User,
+                content: vec![ContentPart::text("What's the current time in UTC?")],
+                controls: Some(Controls {
+                    model_id: None,
+                    locale: None,
+                    reasoning: Some(ReasoningConfig {
+                        effort: Some("low".into()),
+                    }),
+                    error_disclosure: None,
+                    hints: None,
+                }),
+                metadata: None,
+                tags: vec![],
+            };
+            runner.run_turn(input).await.unwrap()
+        }
+    ) else {
         return;
     };
 
-    let runner = InMemoryAgenticLoop::builder()
-        .agent_name("Thinking Time Agent")
-        .system_prompt("Use get_current_time tool when asked about time.")
-        .model(model)
-        .driver_registry(all_providers_registry())
-        .capability(CurrentTimeCapability)
-        .max_iterations(5)
-        .build()
-        .await
-        .unwrap();
-
-    let input = InputMessage {
-        role: MessageRole::User,
-        content: vec![ContentPart::text("What's the current time in UTC?")],
-        controls: Some(Controls {
-            model_id: None,
-            locale: None,
-            reasoning: Some(ReasoningConfig {
-                effort: Some("low".into()),
-            }),
-            error_disclosure: None,
-            hints: None,
-        }),
-        metadata: None,
-        tags: vec![],
-    };
-
-    let result = runner.run_turn(input).await.unwrap();
-
-    skip_if_quota!(result, config.label());
     assert!(result.success, "Turn should succeed: {:?}", result.error);
     assert!(
         result.tool_calls_count > 0,
-        "Should have called get_current_time"
+        "Should have called get_current_time (tool_calls_count={})",
+        result.tool_calls_count
     );
 }
